+++ controls/exampleControl/index.js
@@ -1,0 +1,7 @@
+import ExampleControlComponent from "./components/ExampleControl.vue";
+import ExampleControlStore from "./store/indexExampleControl";
+
+export default {
+    component: ExampleControlComponent,
+    store: ExampleControlStore
+};