+++ controls/exampleControl/store/gettersExampleControl.js
@@ -1,0 +1,6 @@
+import {generateSimpleGetters} from "@shared/js/utils/generators";
+import stateExampleControl from "./stateExampleControl";
+
+export default {
+    ...generateSimpleGetters(stateExampleControl)
+};