+++ controls/exampleControl/store/indexExampleControl.js
@@ -1,0 +1,12 @@
+import actions from "./actionsExampleControl";
+import getters from "./gettersExampleControl";
+import mutations from "./mutationsExampleControl";
+import state from "./stateExampleControl";
+
+export default {
+    namespaced: true,
+    state: {...state},
+    actions,
+    mutations,
+    getters
+};