+++ controls/exampleControl/store/mutationsExampleControl.js
@@ -1,0 +1,6 @@
+import {generateSimpleMutations} from "@shared/js/utils/generators";
+import stateExampleControl from "./stateExampleControl";
+
+export default {
+    ...generateSimpleMutations(stateExampleControl)
+};